CLOUDP-80014: Move Volume/VolumeMount merging logic + tests into merge package (#289)
