+++ pkg/controller/mongodb/replicaset_controller_test.go
@@ -157,13 +157,13 @@ func TestStatefulSet_IsCorrectlyConfigured(t *testing.T) {
 
 	assert.Len(t, sts.Spec.Template.Spec.Containers, 2)
 
-	agentContainer := sts.Spec.Template.Spec.Containers[0]
+	agentContainer := sts.Spec.Template.Spec.Containers[1]
 	assert.Equal(t, agentName, agentContainer.Name)
 	assert.Equal(t, os.Getenv(agentImageEnv), agentContainer.Image)
 	expectedProbe := probes.New(defaultReadiness())
 	assert.True(t, reflect.DeepEqual(&expectedProbe, agentContainer.ReadinessProbe))
 
-	mongodbContainer := sts.Spec.Template.Spec.Containers[1]
+	mongodbContainer := sts.Spec.Template.Spec.Containers[0]
 	assert.Equal(t, mongodbName, mongodbContainer.Name)
 	assert.Equal(t, "repo/mongo:4.2.2", mongodbContainer.Image)
 
@@ -529,8 +529,8 @@ func TestReplicaSet_IsScaledUpToDesiredMembers_WhenFirstCreated(t *testing.T) {
 
 func TestOpenshift_Configuration(t *testing.T) {
 	sts := performReconciliationAndGetStatefulSet(t, "openshift_mdb.yaml")
-	assert.Equal(t, "MANAGED_SECURITY_CONTEXT", sts.Spec.Template.Spec.Containers[0].Env[3].Name)
-	assert.Equal(t, "MANAGED_SECURITY_CONTEXT", sts.Spec.Template.Spec.Containers[1].Env[1].Name)
+	assert.Equal(t, "MANAGED_SECURITY_CONTEXT", sts.Spec.Template.Spec.Containers[1].Env[3].Name)
+	assert.Equal(t, "MANAGED_SECURITY_CONTEXT", sts.Spec.Template.Spec.Containers[0].Env[1].Name)
 }
 
 func TestVolumeClaimTemplates_Configuration(t *testing.T) {
